requirements.txt
FastAPI==0.115.9
uvicorn==0.34.3
python-dotenv==1.1.0

psycopg2-binary==2.9.10
SQLAlchemy>=2.0.0

pydantic>=2.0.0

airbyte==0.31.3
pandas==2.2.3

vanna==0.7.9
chromadb
google-genai
google-cloud-aiplatform==1.96.0
onnxruntime==1.22.0
langchain-google-genai<1.0.0
langchain

torch>=2.2.0,<2.8.0
transformers>=4.35.0,<5.0.0
tokenizers>=0.15.0,<1.0.0
scikit-learn>=1.5.2,<2.0.0
numpy>=1.21.0,<1.27.0

matplotlib==3.9.2

Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
cryptography>=3.4.8
gunicorn>=21.2.0

httpx
requests>=2.25.0
requests-mock==1.11.0

pytz>=2021.1

pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
